[htp] Move the HVX softmax tail and reduction helpers to a header
load_tail_sf, store_tail_sf, reduce_max_sf, reduce_sum_sf and lane0_sf move
verbatim from hvx_softmax_f32.c into hvx_softmax_util.h so the blocked softmax
can reuse them instead of copying. Pure move, no behaviour change.
